Move basic I/O functions into io.ts
